fix(client): forgot to change import after relocating file
